统计模块基本功能，统计历史用户数量和历史订单数量的后台代码

+++ code/scutmall/src/main/java/com/scut/mall/service/OrderService.java
@@ -52,7 +52,7 @@ public interface OrderService {
      * @Param:
      * @return
      */
-    float getMoneySum();
+    double getMoneySum();
     /**
      * create by: Shuo Lin
      * description: 获取用户总数
